feat: minor UI/UX improvements and bug fixes
- Reduce data retention from 30 to 7 days
- Change "Link" to "Cloud" tab with description card
- Add Sign out to tray menu, remove Settings option
- Hide dock icon when signed in (macOS)
- Update app icons from new SVG, separate tray icon
- Add P2P notification card with gradient border
- Fix useEffect infinite loop in password sync
- Fix boolean logic for x=0 edge case in drag-drop
- Use stable React keys instead of array index

+++ desktop/src-tauri/src/mouse_tracking.rs
@@ -36,7 +36,7 @@ pub fn detect_drag(start: &PhysicalPosition<f64>, current: &PhysicalPosition<f64
         use windows::Win32::UI::Input::Mouse::DragDetect;
         use windows::Win32::Foundation::{HWND, POINT};
 
-        let pt = POINT { x: start.x, y: start.y };
+        let pt = POINT { x: start.x as i32, y: start.y as i32 };
         return unsafe { DragDetect(HWND(0), pt).as_bool() };
     }
